% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dem}
\alias{dem}
\title{Digital elevation model}
\format{An object of class \code{geogrid}.}
\source{
Extracted from an elevation model of SE-Germany.
}
\description{
A digital elevation model of a mountainous area
}

