\docType{data}
\name{fieldcapacity}
\alias{fieldcapacity}
\title{Field capacity data}
\format{An object of class \code{geogrid}.}
\source{
  The map corresponds to the digital elevation model in
  \code{dem}.
}
\description{
  A map of the soil's effective field capacity (in
  millimeters)
}
\keyword{datasets}

