% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{landuse}
\alias{landuse}
\title{Land use map}
\format{An object of class \code{geogrid}.}
\source{
The map corresponds to the digital elevation model in \code{dem}.
}
\description{
A land use map (3-digit integer codes)
}

