% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/topocatch.r
\docType{package}
\name{topocatch-package}
\alias{topocatch}
\alias{topocatch-package}
\title{Pre-processing of geo data for hydrological catchment modeling}
\description{
Type \code{help(package="topocatch")} to inspect the package description.
}

