% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xsections_reachPars.r
\name{xs.reachPars}
\alias{xs.reachPars}
\title{Estimation of reach properties for flood routing}
\usage{
xs.reachPars(tabRCH, tabXS, dirXS, dirOUT, american, fun_qMax = function(sqkm)
  {     500 * sqkm^0.43 }, fun_hMax = function(sqkm) {     2 * sqkm^0.2 },
  miniFun = function(dist, diffArea) {     dist/1000 * diffArea },
  almostOne = 0.95, replace = FALSE, plot = FALSE, silent = TRUE)
}
\arguments{
\item{tabRCH}{A data frame listing basic properties of the reaches of interest.
There must be at least the folling columns: 'object' (ID of reach object), 'x1', 'x2', 'y1',
'y2' (coordinates of the reach's end-points), 'upstr_area'
(size of upstream catchment), 'slope' (slope of river bed), 'roughness'
(Manning's n or Strickler's k), 'length' (length of the reach).
Note that a suitable data frame is returned in the
\code{fileAttrRCH} argument of the \code{\link{hydroModelData}} method. The
estimates of the slope in this data frame might need a plausibility check,
however, as there is no guarantee that all values are greater than zero.}

\item{tabXS}{A data frame with the two columns 'datafile' and 'upstr_area'.
Entries in the 1st column should point to files containing the geometry of
individual cross-sections. The 2nd field must contain the catchment sizes
corresponding to the cross-sections. These values are most efficiently
identified in a GIS based on the shape file returned by the
\code{\link{hydroModelData}} method.
The data files referenced in the 1st column must be TAB-separated text files
with (at least) the columns 'x', 'y', 'offset', and 'elevation'. Values in the offset
column represent distances from a point at the bank (usually left bank) and
the values must increase by row. Note that suitable data files are
generated by the \code{\link{xs.extractDEM}} method, for example.}

\item{dirXS}{Name of a directory where the cross-section geometry files
listed in \code{tabXS} reside. This is for the usual case that file
names listed in \code{tabXS} are relative or just basenames.}

\item{dirOUT}{Name of a directory where output files are to be written to.}

\item{american}{A logical value. See the corresponding argument of the
\code{\link{xs.steadyUniformFlow}} method for details.}

\item{fun_qMax}{A function which takes a catchment area as argument and
returns the probable maximum flow rate. Note that the catchment is expected
to in the same units as in the respective fields of \code{tabRCH} and
\code{tabXS}. The function is used to restrict hydraulic computations to
a reasonable range of flow rates. The default function expects the catchment
size in units of square km. It is based on a global study by
Herschy, R. (2001): The world's maximum observed floods, IAHS Publ., vol. 271.}

\item{fun_hMax}{Similar to \code{fun_qMax} but this function must return a
probable maximum flow depth instead of a flow rate. The default yields
values of 2, 5, and 20 meters for catchment sizes of 1, 100, and 100.000
square kilometers, respectively. This function is used to restrict the
evaluation of the cross-section geometry data the approximate range of
interest before fitting analytical functions to the data.}

\item{miniFun}{A function with two formal arguments 'dist' and 'diffArea'
to control the selection of parent cross-sections for a particular reach
from the available data pool.
The 1st argument represents the spatial distance between the
target reach and a potential parent cross-section. The 2nd argument is the
corresponding absolute difference in the upstream catchment areas of the
target reach and the potential parent cross-section, respectively.
The parent cross-sections are selected so as to minimize the value of
\code{miniFun}. The factor of 1000 in the default formulation accounts for
the fact that the distance is in meters (if the x and y coordinates in the
input tables are in meters) while the catchment areas are typically given
in units of square kilometers.}

\item{almostOne}{A number > 0.5 and <= 1. If the interpolation weight for one
of the two 'parent cross-sections' (see below) is greater than this value, this weight
is assumed as 1 and the weight for the other cross-section is set to zero.
Chosing a small value like 0.5 effectively turns the interpolation into a
nearest neighbor approach.}

\item{replace}{If \code{TRUE}, existing files on \code{dirOUT} will be silently replaced.}

\item{plot}{If \code{TRUE}, a graphics files in PDF format is created for each
target location. It can be used to visually assess the quality of the estimate.}

\item{silent}{If \code{TRUE}, some diagnostic messages are printed.}
}
\value{
The function's return value is \code{NULL}. The actual output is
  written to (usually a collection of) text file(s) containing
  all required in information for routing computations in the form of
  look-up tables. These tables are fully self-documenting.
}
\description{
This function computes hydraulic characteristics for river reaches to be
used in flood-routing algorithms of hydrological models. In particular, the
output is useful for flood-routing techniques based on the approach of a
piece-wise linear reservoir. To use this function, cross-section data
must be available at least for some sites in a river basin. These sites are
not necessarily identical with the reaches of interest. See below for details.
}
\note{
To estimate the hydraulic properties of a river reach, its cross-section
  geometry must be know. In river basins, however, such cross-section are
  usually available for selected sites only. Therefore, in the first step,
  the functions 'interpolates' a cross-section for a reach, based on the data
  from two nearby 'parent cross-sections'. These two cross-sections are selected
  from the available data pool so that (1) the cross-sections are nearest to the reach of
  interest and (2) the 1st cross-section has a smaller and the 2nd one has a
  larger upstream catchment compared to the upstream catchment of the reach.
  A cross-section for the reach of interest is then estimated by linearly interpolating
  between the two parent cross-sections using the difference in the upstream catchment size
  as weights.

  This interpolation is, however, \emph{not} applied to the plain
  cross-section's geometry data but to derived characteristics, namely the
  relations \eqn{a(d)} and \eqn{r(d)} where \eqn{d} is the flow depth and
  \eqn{a} and \eqn{r} represent the wet area and the hydraulic radius,
  respectively. First, power models are fitted to represent those functions
  at the parent cross-sections. Second, 'intermediate' power models are
  determined by the above-mentioned interpolation to obtain estimates of
  \eqn{a(d)} and \eqn{r(d)} for the reach of interest. These two functions
  can then be used to compute the reach's hydraulic properties (including a
  rating curve, for example).

  The rationales behind this approach are: (1) Geology may be heteorogenious
  in larger river basins. Hence, it makes sense to assume that two nearby
  cross-sections are likely to be more 'similar' in terms of its properties
  than two cross-sections being located far away from each other. This is
  why the parent cross-sections are selected considering a minimum-distance
  criterion (see argument \code{miniFun}).
  (2) The cross-section geometry (namely the flow capacity) is assumed to be
  corellated with the size of the upstream catchment. Therefore, the
  parent cross-section whose catchment size is closer to the catchment size
  of the reach of interest is assigned a greater weight in the interpolation.
}
\author{
David Kneis \email{david.kneis@uni-potsdam.de} while being at
  IIT Kharagpur, India
}

