% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/private_common_geogrid.r
\name{is.geogrid}
\alias{is.geogrid}
\title{Test whether an object is of class geogrid}
\usage{
is.geogrid(x)
}
\arguments{
\item{x}{The object to be tested.}
}
\value{
A logical value.
}
\description{
A valid geogrid object is a of class geogrid and consists of a list with
seven named components 'ncols', 'nrows', 'xllcorner', 'yllcorner',
'cellsize', 'nodata_value', and 'matrix'.
}
\author{
David Kneis \email{david.kneis@uni-potsdam.de}
}

